requirements.txt
# Main dependencies for the script
requests
apscheduler
elasticsearch
SQLAlchemy
CMRESHandler

# Dependencies for Telegram bot
python-telegram-bot==12.0.0b1